+++ R/class_core_parameter_set.R
@@ -13,8 +13,8 @@
 ## |
 ## |  Contact us for information about our services: info@rpact.com
 ## |
-## |  File version: $Revision: 8615 $
-## |  Last changed: $Date: 2025-03-17 16:43:46 +0100 (Mo, 17 Mrz 2025) $
+## |  File version: $Revision: 8616 $
+## |  Last changed: $Date: 2025-03-18 16:07:37 +0100 (Di, 18 Mrz 2025) $
 ## |  Last changed by: $Author: pahlke $
 ## |
 
@@ -1335,9 +1335,10 @@ ParameterSet <- R6::R6Class("ParameterSet",
     }
 
     if (!inherits(parameterSet, "AccrualTime")) {
-        accrualTime <- parameterSet[["accrualTime"]]
+        accrualTime <- parameterSet[[".accrualTime"]]
         if (!is.null(accrualTime) && length(accrualTime) > 1) {
-            parametersToIgnore <- c(parametersToIgnore, c("accrualTime", "accrualIntensity"))
+            parametersToIgnore <- c(parametersToIgnore, 
+                c("accrualTime", "accrualIntensity", "accrualIntensityRelative"))
         }
     }